requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.6.1
    # via
    #   starlette
    #   watchfiles
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   typer
    #   uvicorn
fastapi==0.85.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
httptools==0.5.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
numpy==1.23.3
    # via
    #   pandas
    #   pyarrow
pandas==1.5.0
    # via -r requirements.in
pyarrow==9.0.0
    # via -r requirements.in
pydantic==1.10.2
    # via
    #   -r requirements.in
    #   fastapi
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.21.0
    # via uvicorn
pytz==2022.4
    # via pandas
pyyaml==6.0
    # via uvicorn
requests==2.28.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via anyio
starlette==0.20.4
    # via fastapi
typer==0.6.1
    # via -r requirements.in
typing-extensions==4.3.0
    # via pydantic
urllib3==1.26.12
    # via requests
uvicorn[standard]==0.18.3
    # via -r requirements.in
uvloop==0.17.0
    # via uvicorn
watchfiles==0.17.0
    # via uvicorn
websockets==10.3
    # via uvicorn
